+++ core/src/main/kotlin/dev/zieger/utils/delegates/OnChangedScope.kt
@@ -2,12 +2,6 @@
 
 package dev.zieger.utils.delegates
 
-
-/**
- * Same as [IOnChangedScope2] but with [Any]? as parent type.
- */
-interface IOnChangedScope<out T : Any?> : IOnChangedScope2<Any?, T>
-
 /**
  * Scope that describes the change of any property of type [T] hold by a parent of type [P].
  *
@@ -17,8 +11,8 @@ interface IOnChangedScope<out T : Any?> : IOnChangedScope2<Any?, T>
  * @property previousValues all previous values of the property
  * @property clearPreviousValues Lambda to clear the previous values
  */
-interface IOnChangedScope2<P : Any?, out T : Any?> {
-    val value: @UnsafeVariance T
+interface IOnChangedScope2<out P : Any?, out T : Any?> {
+    val value: T
     val thisRef: P?
     val previousValue: T?
     val previousValues: List<T?>
@@ -26,26 +20,18 @@ interface IOnChangedScope2<P : Any?, out T : Any?> {
     val isInitialNotification: Boolean
 }
 
-/**
- * Same as [OnChangedScope2] but with [Any]? as parent type.
- */
-data class OnChangedScope<out T : Any?>(
-    override val value: @UnsafeVariance T,
-    override val thisRef: Any?,
-    override val previousValue: T?,
-    override val previousValues: List<T?>,
-    override val clearPreviousValues: () -> Unit,
-    override val isInitialNotification: Boolean = false
-) : IOnChangedScope<T>
+typealias IOnChangedScope<T> = IOnChangedScope2<Any?, T>
 
 /**
  * Simple implementation of [IOnChangedScope2].
  */
 data class OnChangedScope2<P : Any?, out T : Any?>(
-    override val value: @UnsafeVariance T,
+    override val value: T,
     override val thisRef: P?,
     override val previousValue: T?,
     override val previousValues: List<T?>,
     override val clearPreviousValues: () -> Unit,
     override val isInitialNotification: Boolean = false
 ) : IOnChangedScope2<P, T>
+
+typealias OnChangedScope<T> = OnChangedScope2<Any?, T>