+++ core/src/main/kotlin/dev/zieger/utils/time/TimeEx.kt
@@ -1,19 +1,23 @@
 package dev.zieger.utils.time
 
-import dev.zieger.utils.misc.FiFo
 import dev.zieger.utils.time.base.TimeUnit
 import dev.zieger.utils.time.base.TimeUnit.MS
 import dev.zieger.utils.time.base.toMillis
 import dev.zieger.utils.time.duration.IDurationHolder
 import java.util.*
-import kotlin.collections.HashMap
+import java.util.concurrent.atomic.AtomicLong
 
 open class TimeEx(
     override val millis: Long = System.currentTimeMillis(),
     override val zone: TimeZone = TimeZone.getDefault()
 ) : ITimeEx {
 
-    companion object : TimeParseHelper()
+    companion object : TimeParseHelper() {
+        private var lastId = AtomicLong(-1L)
+        private val newId get() = lastId.incrementAndGet()
+    }
+
+    override val id: Long = newId
 
     constructor(value: Number, timeUnit: TimeUnit = MS, timeZone: TimeZone = TimeZone.getDefault()) :
             this(value.toLong().toMillis(timeUnit), timeZone)
@@ -23,9 +27,6 @@ open class TimeEx(
 
     constructor(date: Date, timeZone: TimeZone = TimeZone.getDefault()) :
             this(date.time, timeZone)
-
-    @Suppress("ReplaceWithEnumMap")
-    override val timeUnitLengthMap: HashMap<Long, FiFo<Int>> = HashMap()
 
     override fun toString() = formatTime(DateFormat.COMPLETE)
     override fun equals(other: Any?) = millis == (other as? ITimeEx)?.millis