+++ core/src/test/kotlin/dev/zieger/utils/delegates/OnChangedTest.kt
@@ -81,16 +81,20 @@ class OnChangedTest : AnnotationSpec() {
 
             var propertyValue: () -> TestValueContainer = { TestValueContainer() }
             var testProperty by OnChanged(
-                TestValueContainer(), storePreviousValues, notifyForInitial,
-                notifyOnChangedOnly, scope, mutex, { newVeto },
-                { v ->
-                    if (suspendedListener && scope != null)
-                        onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
-                },
-                { v ->
-                    if (!suspendedListener || scope == null)
-                        onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
-                })
+                OnChangedParams(
+                    TestValueContainer(), storeRecentValues = storePreviousValues,
+                    notifyForInitial = notifyForInitial,
+                    notifyOnChangedValueOnly = notifyOnChangedOnly,
+                    scope = scope, mutex = mutex, veto = { newVeto },
+                    onChangedS = { v ->
+                        if (suspendedListener && scope != null)
+                            onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
+                    },
+                    onChanged = { v ->
+                        if (!suspendedListener || scope == null)
+                            onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
+                    })
+            )
             propertyValue = { testProperty }
 
             repeat(1000) { i ->
@@ -117,16 +121,21 @@ class OnChangedTest : AnnotationSpec() {
             val results = Channel<OnChangedResults>(Channel.UNLIMITED)
 
             var propertyValue: () -> TestValueContainer = { TestValueContainer() }
-            var testProperty by OnChanged2(TestValueContainer(), storePreviousValues, notifyForInitial,
-                notifyOnChangedOnly, scope, mutex, { newVeto },
-                { v ->
-                    if (suspendedListener && scope != null)
-                        onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
-                },
-                { v ->
-                    if (!suspendedListener || scope == null)
-                        onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
-                })
+            var testProperty by OnChanged2(
+                OnChangedParams2(
+                    TestValueContainer(), storeRecentValues = storePreviousValues,
+                    notifyForInitial = notifyForInitial,
+                    notifyOnChangedValueOnly = notifyOnChangedOnly,
+                    scope = scope, mutex = mutex, veto = { newVeto },
+                    onChangedS = { v ->
+                        if (suspendedListener && scope != null)
+                            onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
+                    },
+                    onChanged = { v ->
+                        if (!suspendedListener || scope == null)
+                            onChangedListenerBlock(newClearCache, results, newValue, propertyValue, newVeto)
+                    })
+            )
             propertyValue = { testProperty }
 
             repeat(1000) { i ->